Setup Hive notifications
