+++ mobile/lib/screens/profile/profile_edit_page.dart
@@ -12,13 +12,13 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:image_picker/image_picker.dart';
 
-import '../auth/change_email_screen.dart';
-import '../auth/change_phone_screen.dart';
-import '../models/enum_constants.dart';
-import '../services/app_service.dart';
-import '../services/firebase_service.dart';
-import '../themes/light_theme.dart';
-import 'home_page.dart';
+import '../../auth/change_email_screen.dart';
+import '../../auth/change_phone_screen.dart';
+import '../../models/enum_constants.dart';
+import '../../services/app_service.dart';
+import '../../services/firebase_service.dart';
+import '../../themes/light_theme.dart';
+import '../home_page.dart';
 
 class ProfileEditPage extends StatefulWidget {
   final UserDetails userDetails;
