+++ mobile/lib/screens/web_view_page.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
-import 'package:app/services/app_service.dart';
 import 'package:app/utils/dialogs.dart';
 import 'package:app/utils/extensions.dart';
+import 'package:app/utils/network.dart';
 import 'package:flutter/material.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 
@@ -23,7 +23,6 @@ class WebViewScreen extends StatefulWidget {
 class _WebViewScreenState extends State<WebViewScreen> {
   final controller = Completer<WebViewController>();
   var loadingPercentage = 0;
-  final AppService _appService = AppService();
 
   @override
   Widget build(BuildContext context) {
@@ -71,7 +70,7 @@ class _WebViewScreenState extends State<WebViewScreen> {
   @override
   void initState() {
     super.initState();
-    _appService.isConnected(context);
+    checkNetworkConnection(context);
   }
 }
