+++ mobile/lib/services/hive_service.dart
@@ -1,12 +1,15 @@
 import 'package:hive_flutter/hive_flutter.dart';
 
 import '../constants/config.dart';
+import '../models/enum_constants.dart';
 import '../models/notification.dart';
 
 class HiveService {
   static Future<void> initialize() async {
     await Hive.initFlutter();
-    Hive.registerAdapter(AppNotificationAdapter());
+    Hive
+      ..registerAdapter(AppNotificationAdapter())
+      ..registerAdapter(AppNotificationTypeAdapter());
 
     await Hive.openBox<AppNotification>(HiveBox.appNotifications);
   }
