+++ mobile/lib/utils/network.dart
@@ -1,6 +1,11 @@
 import 'dart:io';
 
-Future<bool> hasFirebaseConnection() async {
+import 'package:flutter/material.dart';
+
+import '../constants/config.dart';
+import 'dialogs.dart';
+
+Future<bool> hasNetworkConnection() async {
   try {
     final result = await InternetAddress.lookup('firebase.google.com');
     if (result.isNotEmpty && result[0].rawAddress.isNotEmpty) {
@@ -8,4 +13,12 @@ Future<bool> hasFirebaseConnection() async {
     }
   } on Exception catch (_) {}
   return false;
+}
+
+Future<bool> checkNetworkConnection(BuildContext buildContext) async {
+  var hasConnection = await hasNetworkConnection();
+  if (!hasConnection) {
+    await showSnackBar(buildContext, Config.connectionErrorMessage);
+  }
+  return hasConnection;
 }
